client_api/requirements.txt
fastapi==0.95.1
pandas==1.4.1
uvicorn==0.21.1
psycopg2-binary==2.9.5
python-dotenv==0.20.0
gunicorn==20.1.0 